requirements.txt
Django==2.2
psycopg2-binary==2.8.5
gunicorn==19.9.0
django-safedelete==0.5.2
django_extensions==2.1.4
djangorestframework==3.9.0
django-guardian==2.3.0
django-model-utils==3.1.2
django-multitenant==2.0.9
django-cors-headers==2.5.2
requests